Move api check to MyApp.
